+++ CHANGELOG.md
@@ -4,6 +4,11 @@ The change log documents features and breaking changes for each version.
 While RESTless reaches a 1.0 release and more feedback and testing continues, the API is subject to change.  
 To keep the library small, API depreciation warnings may not be added.
 
+### 0.4.2
+
+- YUIDoc documentation 
+- Bug fixes 
+
 ### 0.4.1
 
 - Register via `Ember.libraries`